+++ Packages/Tests/CLICommandTests/FetchTests/CrawlTests.swift
@@ -48,6 +48,8 @@ struct WebCrawlTests {
             htmlParser: LiveTestHTMLParserStrategy(),
             appleJSONParser: LiveTestAppleJSONParserStrategy(),
             priorityPackageStrategy: LiveTestPriorityPackageStrategy(),
+
+            fetcherFactory: Crawler.NoopHTTPFetcherFactory(),
             logger: Logging.NoopRecording()
         )
         let stats = try await crawler.crawl()
@@ -116,6 +118,8 @@ struct WebCrawlTests {
             htmlParser: LiveTestHTMLParserStrategy(),
             appleJSONParser: LiveTestAppleJSONParserStrategy(),
             priorityPackageStrategy: LiveTestPriorityPackageStrategy(),
+
+            fetcherFactory: Crawler.NoopHTTPFetcherFactory(),
             logger: Logging.NoopRecording()
         )
         let stats1 = try await crawler1.crawl()
@@ -127,6 +131,8 @@ struct WebCrawlTests {
             htmlParser: LiveTestHTMLParserStrategy(),
             appleJSONParser: LiveTestAppleJSONParserStrategy(),
             priorityPackageStrategy: LiveTestPriorityPackageStrategy(),
+
+            fetcherFactory: Crawler.NoopHTTPFetcherFactory(),
             logger: Logging.NoopRecording()
         )
         let stats2 = try await crawler2.crawl()